08-YunPian/requirements.txt
opencv_python==4.0.0.21
requests==2.22.0
PyExecJS==1.5.1
